refactor: rename server-roles.utils.ts file
